refactor(hooks): useAppInitialize 초기화 로직 internal 모듈 분할 (Phase 2)
- appInitializeImports: 동적 import 폴백·describeError (import 문자열 원문 유지)
- appInitializeProfile: 프로필 로드·클레임 동기화 판정 헬퍼
- appInitializeSession: bootstrap·resolveSession·reconcileSessionFromServer + 타입
- 진입점 re-export(resolveSession/reconcileSessionFromServer/waitForInitialAuthUser/OfflineBootstrapState/default)로 소비자·테스트 호환
- 초기화 순서·시그니처 불변, 코드 이동만 (935→314줄, tsc 0, test 4286 pass)
